Updates UK bus data source URL
Updates the UK bus data source URL to bustimes.org.

The previous URL, ukbuses.org, appears to be outdated or no longer functioning as expected, so the application is updated to use the new bustimes.org endpoint. This ensures that the application continues to receive current and accurate data for UK bus stops.

Removes the addition of `no_inactive` and `contains_delete` params to the base URL, as they are no longer valid in the new API.